backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
jinja2
pytest
httpx
python-dotenv
flake8

